requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.6.4
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
